requirements.txt
tensorflow-macos==2.13.0
opencv-python==4.10.0.84
numpy==1.22.4
matplotlib==3.5.1
tqdm